+++ CHANGELOG.md
@@ -24,15 +24,13 @@ Semantic versioning in our case means:
 This version introduces `wps` CLI tool.
 - `wps explain <code>` command can be used to access WPS violation docs (same as on website) without internet access
 
-
-## 1.0.1 WIP
-
 ### Bugfixes
 
 - Fixes `WPS115` false-positive on `Enum` attributes, #3238
 - Removes duplicated `WPS312`, #3239
 - Fixes `WPS432`, now it shows literal num, #1402
 - Fixes `WPS605` false-positive on `@staticmethod`, #3292
+
 
 ## 1.0.0